% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRSpss.R
\name{molgenis.import.csv.metadata}
\alias{molgenis.import.csv.metadata}
\title{Import the metadata from the SPSS-file}
\usage{
molgenis.import.csv.metadata()
}
\value{
the importer job url
}
\description{
The import opf metadata from the SPSS-file is performed by the MOLGNIS importers.
You have to generate a CSV file and then import it with this function.
}
\details{
Default file-name importing for metadata is attributes.csv
}
