% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRSpss.R
\name{molgenis.import.spss}
\alias{molgenis.import.spss}
\title{Import the SPSS-file}
\usage{
molgenis.import.spss(spss.file)
}
\arguments{
\item{spss.file}{SPSS-file to import}

\item{molgenis.host}{MOLGENIS host you want to import in}
}
\description{
You can import an SPSS file in MOLGENIS via the importer.
You have to have a secure connection to import a file into MOLGENIS.
}
