% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRSpss.R
\name{molgenis.job.status}
\alias{molgenis.job.status}
\title{Poll for scheduled job in MOLGENIS}
\usage{
molgenis.job.status(job.url)
}
\arguments{
\item{job.url}{job URL which can be used to check the status}
}
\value{
the job status (FINISHED, RUNNING or FAILED)
}
\description{
You have to check if the status of the job is FINISHED
A job can have 3 possible statusses
- FINISHED
- RUNNING
- FAILED
}
