% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molgenisRSpss.R
\name{molgenis.spss.import}
\alias{molgenis.spss.import}
\title{Import the SPSS-file}
\usage{
molgenis.spss.import(molgenis.host, molgenis.token, spss.file)
}
\arguments{
\item{molgenis.host}{molgenis host}

\item{molgenis.token}{molgenis token}

\item{spss.file}{SPSS-file to import}
}
\description{
You can import an SPSS file in MOLGENIS via the importer.
You have to have a secure connection to import a file into MOLGENIS.
The MOLGENIS R API can be used to setup a connection.
molgenis.login(#host#, #username#, #password#) can be used.
}
